feat: add a button to create new spots
